# docs/swagger.yaml
basePath: /v1
definitions:
...
  github_com_pickle_pw_monolith_db_sqlc.ReactionSource:
    enum:
    - unicode_emoji
    - 7tv
    - custom
    type: string
    x-enum-varnames:
    - ReactionSourceUnicodeEmoji
    - ReactionSource7tv
    - ReactionSourceCustom
  github_com_pickle_pw_monolith_internal_models.ReactionStack:
    properties:
      count:
        type: integer
      emoteId:
        type: string
      noteId:
        type: string
      source:
        $ref: '#/definitions/github_com_pickle_pw_monolith_db_sqlc.ReactionSource'
      userReacted:
        type: boolean
    type: object
...
  github_com_pickle_pw_monolith_internal_types.BatchNoteReactionsRes:
    properties:
      noteId:
        type: string
      reactions:
        items:
          $ref: '#/definitions/github_com_pickle_pw_monolith_internal_models.ReactionStack'
        type: array
    type: object
...
  github_com_pickle_pw_monolith_internal_types.ReactionReq:
    properties:
      emoteId:
        type: string
      source:
        type: string
    type: object
...
host: api.pickle.pw
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a Pickle server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Pickle API
  version: "1.0"
paths:
...
  /v1/users/{userId}/content-notes/{category}:
    get:
      consumes:
      - application/json
      description: Get content notes sorted by receiver link
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Cursor
        in: query
        name: cursor
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Column
        in: query
        name: column
        type: string
      - description: Direction
        in: query
        name: direction
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items: {}
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get content notes sorted by receiver link
      tags:
      - content-notes
    post:
      consumes:
      - application/json
      description: Create a content note
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Content note request
        in: body
        name: contentNoteReq
        required: true
        schema: {}
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create a content note
      tags:
      - content-notes
...
  /v1/users/{userId}/content-notes/{category}/{noteId}/reactions:
    delete:
      consumes:
      - application/json
      description: Remove content note reaction
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Note ID
        in: path
        name: noteId
        required: true
        type: string
      - description: Reaction request
        in: body
        name: reactionReq
        required: true
        schema:
          $ref: '#/definitions/github_com_pickle_pw_monolith_internal_types.ReactionReq'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Remove content note reaction
      tags:
      - content-notes
    get:
      consumes:
      - application/json
      description: Get content note reactions
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Note ID
        in: path
        name: noteId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_pickle_pw_monolith_internal_models.ReactionStack'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get content note reactions
      tags:
      - content-notes
    post:
      consumes:
      - application/json
      description: Add content note reaction
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Note ID
        in: path
        name: noteId
        required: true
        type: string
      - description: Reaction request
        in: body
        name: reactionReq
        required: true
        schema:
          $ref: '#/definitions/github_com_pickle_pw_monolith_internal_types.ReactionReq'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Add content note reaction
      tags:
      - content-notes
  /v1/users/{userId}/content-notes/{category}/by-content-id/{contentId}:
    get:
      consumes:
      - application/json
      description: Search content note by content ID
      parameters:
      - description: Content ID
        in: path
        name: contentId
        required: true
        type: string
      - description: Category
        in: path
        name: category
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Search content note by content ID
      tags:
      - content-notes
  /v1/users/{userId}/content-notes/reactions:
    get:
      consumes:
      - application/json
      description: Get batch content note reactions
      parameters:
      - description: Note IDs
        in: query
        name: noteIds
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_pickle_pw_monolith_internal_types.BatchNoteReactionsRes'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get batch content note reactions
      tags:
      - content-notes
...
swagger: "2.0"